requirements.txt
# Core dependencies
google-genai>=1.0.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0

# Testing
pytest>=8.0.0
pytest-mock>=3.14.0

# UI
streamlit>=1.32.0

# PoC Analysis (Iteration-05)
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.1.0
